+++ src/main/java/dcc/cc3002/king/cards/CardType.java
@@ -1,5 +1,0 @@
-package dcc.cc3002.king.cards;
-
-public enum CardType {
-  MONSTER, MAGIC
-}